# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Daftar Properti API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tele-app/listings": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Get listing items",
                "description": "Returns listing items",
                "operationId": "listings.index",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search listing by keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "collection",
                        "in": "query",
                        "description": "If set to true, it will only return user's collection",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "price[min]",
                        "in": "query",
                        "description": "Minimum price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price[max]",
                        "in": "query",
                        "description": "Maximum price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "propertyType",
                        "in": "query",
                        "description": "Property type",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/PropertyType"
                        }
                    },
                    {
                        "name": "bedroomCount",
                        "in": "query",
                        "description": "Bedroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bedroomCount[min]",
                        "in": "query",
                        "description": "Minimum Bedroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bedroomCount[max]",
                        "in": "query",
                        "description": "Maximum Bedroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathroomCount",
                        "in": "query",
                        "description": "Bathroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathroomCount[min]",
                        "in": "query",
                        "description": "Minimum Bathroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathroomCount[max]",
                        "in": "query",
                        "description": "Maximum Bathroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lotSize[min]",
                        "in": "query",
                        "description": "Minimum lot size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lotSize[max]",
                        "in": "query",
                        "description": "Maximum lot size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "buildingSize[min]",
                        "in": "query",
                        "description": "Minimum building size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "buildingSize[max]",
                        "in": "query",
                        "description": "Maximum building size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ownership",
                        "in": "query",
                        "description": "Ownership",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/PropertyOwnership"
                        }
                    },
                    {
                        "name": "carCount",
                        "in": "query",
                        "description": "Car count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "carCount[min]",
                        "in": "query",
                        "description": "Minimum Car count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "carCount[max]",
                        "in": "query",
                        "description": "Maximum Car count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "electricPower",
                        "in": "query",
                        "description": "Electric Power",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort By",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/ListingSort"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order By",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "listings": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Listing"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Listings"
                ],
                "summary": "Create listing",
                "operationId": "listings.create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ListingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Listing"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tele-app/listings/{id}": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Get listing by id",
                "operationId": "listings.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Listing"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Listing not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Listing not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Listings"
                ],
                "summary": "Update listing",
                "operationId": "listings.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ListingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Listing"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Listing not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Listing not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Listings"
                ],
                "summary": "Delete listing",
                "operationId": "listings.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Listing deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ListingSort": {
                "description": "Sort Listing By",
                "type": "string",
                "enum": [
                    "price",
                    "bedroomCount",
                    "bathroomCount",
                    "lotSize",
                    "buildingSize"
                ],
                "example": "price"
            },
...
            "PropertyOwnership": {
                "description": "Property ownership/certificate",
                "type": "string",
                "enum": [
                    "unknown",
                    "shm",
                    "hgb",
                    "strata",
                    "girik"
                ],
                "example": "shm"
            },
            "PropertyType": {
                "description": "Property type",
                "type": "string",
                "enum": [
                    "unknown",
                    "house",
                    "apartment",
                    "warehouse",
                    "shophouse",
                    "land",
                    "villa"
                ],
                "example": "house"
            },
...
        }
    }
}